requirements.txt
# Core framework
fastapi>=0.115.0
pydantic>=2.4.0

# Telegram Bot
aiogram>=3.22.0

# AI - OpenRouter (OpenAI-compatible API for LLM + Embeddings)
openai>=1.54.3
tenacity>=8.2.0

# AI - LangChain/LangGraph Agent (via OpenAI-compatible)
langchain-openai>=0.2.0
langgraph>=0.2.0
langchain-core>=0.3.81

# Database & Cache
supabase>=2.27.2
upstash-redis==1.5.0

# Payments (lazy loaded)
stripe>=10.0.0

# QStash
qstash>=3.0.0

# HTTP client
httpx>=0.27.0

# Testing
pytest>=7.4.0

# Security - Aikido Zen (Runtime Firewall)
aikido_zen>=1.0.0

# Observability - Digma (Code Analysis: N+1, Chatty Logic, Dead Code)
# Optional: Install only if DIGMA_COLLECTOR_URL and DIGMA_API_KEY are set
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0
opentelemetry-exporter-otlp>=1.21.0

# REMOVED (unused):
# uvicorn - Vercel doesn't need it
# AaioAPI - not imported anywhere
# python-i18n - using custom src/i18n
# python-dotenv - only for local scripts
